refactor: extract filter query logic
